fix: implemnt hook emailsend whit tansktack query
