+++ apps/dashboard/src/pages/CoordinatorEmail.tsx
@@ -1,24 +1,18 @@
 import { useRef, useState } from "react";
-import toast from "react-hot-toast";
-
-type GraphFileAttachment = {
-  "@odata.type": "#microsoft.graph.fileAttachment";
-  name: string;
-  contentType: string;
-  contentBytes: string; // base64
-};
+import {
+  useSendMail,
+  MAX_FILES,
+  MAX_FILE_BYTES,
+  MAX_TOTAL_BYTES,
+} from "../hooks/useSendMail";
 
 export default function CoordinatorEmail() {
   const [to, setTo] = useState("");
   const [subject, setSubject] = useState("");
   const [body, setBody] = useState("");
-  const [isSending, setIsSending] = useState(false);
   const [files, setFiles] = useState<File[]>([]);
   const fileInputRef = useRef<HTMLInputElement | null>(null);
-
-  const MAX_FILE_BYTES = 3 * 1024 * 1024;
-  const MAX_TOTAL_BYTES = 10 * 1024 * 1024;
-  const MAX_FILES = 5;
+  const { sendMail, isSending } = useSendMail();
 
   const humanSize = (bytes: number) => {
     if (bytes < 1024) return `${bytes} B`;
@@ -32,28 +26,15 @@ export default function CoordinatorEmail() {
 
     const capacity = MAX_FILES - files.length;
     if (capacity <= 0) {
-      toast.error(`Máximo ${MAX_FILES} archivos`);
       if (fileInputRef.current) fileInputRef.current.value = "";
       return;
     }
 
     const toAdd = selected.slice(0, capacity);
-
-    const validBySize = toAdd.filter((f) => {
-      const ok = f.size <= MAX_FILE_BYTES;
-      if (!ok)
-        toast.error(`Archivo muy grande: ${f.name} (${humanSize(f.size)})`);
-      return ok;
-    });
-
+    const validBySize = toAdd.filter((f) => f.size <= MAX_FILE_BYTES);
     const currentTotal = files.reduce((acc, f) => acc + f.size, 0);
     const incomingTotal = validBySize.reduce((acc, f) => acc + f.size, 0);
     if (currentTotal + incomingTotal > MAX_TOTAL_BYTES) {
-      toast.error(
-        `Límite total excedido (${humanSize(
-          currentTotal + incomingTotal,
-        )} > ${humanSize(MAX_TOTAL_BYTES)})`,
-      );
       if (fileInputRef.current) fileInputRef.current.value = "";
       return;
     }
@@ -62,118 +43,17 @@ export default function CoordinatorEmail() {
     if (fileInputRef.current) fileInputRef.current.value = "";
   };
 
-  const fileToBase64 = (f: File): Promise<string> =>
-    new Promise((resolve, reject) => {
-      const reader = new FileReader();
-      reader.onload = () => {
-        const result = reader.result as string;
-        const base64 = result.split(",")[1] ?? "";
-        resolve(base64);
-      };
-      reader.onerror = (err) => reject(err);
-      reader.readAsDataURL(f);
-    });
-
   const handleSend = async () => {
-    if (!to.trim() || !subject.trim() || !body.trim()) {
-      toast.error("Completa destinatario, asunto y mensaje");
-      return;
-    }
-
     try {
-      const mailToken = sessionStorage.getItem("mailToken");
-      if (!mailToken) {
-        toast.error("No se encontró mailToken en la sesión");
-        return;
-      }
-
-      setIsSending(true);
-      const loadingId = toast.loading("Enviando correo...");
-
-      let attachments: GraphFileAttachment[] | undefined = undefined;
-      if (files.length) {
-        if (files.length > MAX_FILES) {
-          toast.dismiss(loadingId);
-          setIsSending(false);
-          toast.error(`Máximo ${MAX_FILES} archivos`);
-          return;
-        }
-        const totalBytes = files.reduce((acc, f) => acc + f.size, 0);
-        if (totalBytes > MAX_TOTAL_BYTES) {
-          toast.dismiss(loadingId);
-          setIsSending(false);
-          toast.error(
-            `Límite total excedido (${humanSize(totalBytes)} > ${humanSize(
-              MAX_TOTAL_BYTES,
-            )})`,
-          );
-          return;
-        }
-        for (const f of files) {
-          if (f.size > MAX_FILE_BYTES) {
-            toast.dismiss(loadingId);
-            setIsSending(false);
-            toast.error(`Archivo muy grande: ${f.name}`);
-            return;
-          }
-        }
-        const converted: GraphFileAttachment[] = [];
-        for (const f of files) {
-          const contentBytes = await fileToBase64(f);
-          converted.push({
-            "@odata.type": "#microsoft.graph.fileAttachment",
-            name: f.name,
-            contentType: f.type || "application/octet-stream",
-            contentBytes,
-          });
-        }
-        attachments = converted;
-      }
-
-      const res = await fetch("https://graph.microsoft.com/v1.0/me/sendMail", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${mailToken}`,
-        },
-        body: JSON.stringify({
-          message: {
-            subject,
-            body: {
-              contentType: "HTML",
-              content: body,
-            },
-            toRecipients: [
-              {
-                emailAddress: {
-                  address: to,
-                },
-              },
-            ],
-            ...(attachments ? { attachments } : {}),
-          },
-          saveToSentItems: "true",
-        }),
-      });
-
-      toast.dismiss(loadingId);
-      setIsSending(false);
-
-      if (!res.ok) {
-        const txt = await res.text().catch(() => "");
-        throw new Error(txt || "Error al enviar el correo");
-      }
-
-      toast.success("Mensaje enviado con éxito");
+      await sendMail({ to, subject, body, files });
       setTo("");
       setSubject("");
       setBody("");
       setFiles([]);
       if (fileInputRef.current) fileInputRef.current.value = "";
-    } catch (err) {
-      setIsSending(false);
-      const msg = err instanceof Error ? err.message : String(err);
-      toast.error(msg);
+    } catch (error) {
+      // Los errores ya están manejados por el hook con toasts
+      console.error("Error sending mail:", error);
     }
   };
